[universeII] don't change the PCI BAR values assigned during system boot
The base address registers (BAR) in the PCI configuration space of all
PCI devices are set to valid, unique values during system boot (either
by the bios or the operating system).
There is no need to change these values, we can just read them, request
the kernel to use them, and then ioremap them.

Changing the universe BAR to 0xA0000000, as was done before, and not
reseving the address, is dangerous (although so far this has always
worked).