[vmelib] always use `unsigned long` as parameter for ioctl calls
The ioctl handler in the universeII driver (`universeII_ioctl()`) always
reads an `unsigned long` from userspace. (And for some calls we actually
need this, because it needs to fit a pointer!)
We should therefore allways also pass an `unsigned long` to the `ioctl`
call, event for commands that only need an int.
This was not a real issue in the past, because in the ioctl handler the
bits that are actually used are masked out (or the value is implicitly
cast to `unsigned int`). Since we are only using little endian systems,
this still did work out fine.
Nevertheless, it is much safer to always pass an `unsigned long`, to
ensure that the ioctl handler never reads invalid data into the upper
bytes of the `arg` parameter on 64bit systems, but they are always
properly initialized to 0.